+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -4,6 +4,8 @@ import java.util.List;
 import javax.validation.Valid;
 import javax.validation.constraints.Positive;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -22,6 +24,7 @@ import lombok.extern.slf4j.Slf4j;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
 
+@Validated
 @RestController
 @Slf4j
 @RequestMapping("/films")
@@ -29,8 +32,9 @@ import ru.yandex.practicum.filmorate.service.FilmService;
 public class FilmController {
     final FilmService filmService;
 
-    public FilmController(FilmService service) {
-        this.filmService = service;
+    @Autowired
+    public FilmController(FilmService filmService) {
+        this.filmService = filmService;
     }
 
     @PostMapping
@@ -40,7 +44,7 @@ public class FilmController {
     }
 
     @PutMapping
-    public Film updateFilm(@Valid @RequestBody Film film) {
+    public Film updateFilm(@Valid @RequestBody Film film) throws JsonProcessingException {
         log.info("Получен запрос на обновление фильма id={}", film.getId());
         return filmService.updateFilm(film);
     }
