+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -2,20 +2,14 @@ package ru.yandex.practicum.filmorate.storage.user;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 
-import java.time.LocalDate;
 import java.util.*;
 import java.util.stream.Collectors;
 
 @Slf4j
 @Component
 public class InMemoryUserStorage implements UserStorage {
-    private static final ObjectMapper mapper = new ObjectMapper().findAndRegisterModules();
     private final Map<Integer, User> users = new HashMap<>();
 
     public Map<Integer, User> getUsers() {
@@ -91,25 +85,5 @@ public class InMemoryUserStorage implements UserStorage {
 
     public boolean areTheseUsersFriends(Integer userId, Integer friendId) {
         return (users.get(userId).getSetOfFriends().contains(friendId));
-    }
-
-    public static boolean isUserValid(User user) throws JsonProcessingException {
-        return isLoginValid(user) && isBirthdayValid(user);
-    }
-
-    private static boolean isBirthdayValid(User user) throws JsonProcessingException {
-        if (user.getBirthday().isAfter(LocalDate.now())) {
-            log.warn("Дата рождения {} не может быть в будущем", mapper.writeValueAsString(user.getBirthday()));
-            throw new ValidationException("Дата рождения " + user.getBirthday() + "не может быть из будущего");
-        }
-        return true;
-    }
-
-    private static boolean isLoginValid(User user) {
-        if (user.getLogin().contains(" ")) {
-            log.warn("Логин {} содержит пробелы", user.getLogin());
-            throw new ValidationException("Логин " + user.getLogin() + " содержит пробелы");
-        }
-        return true;
     }
 }
